+++ src/components/EditPanel/EditSidebar.jsx
@@ -1,0 +1,38 @@
+import CustomButton from "../CustomButton";
+import PropTypes from "prop-types";
+/*
+- Section for switching tabs, which affects the content of the form-section
+*/
+function EditSideBar({ tabType, onTabChange }) {
+	const btnClassList = ["sidebar-btn", "button-shrink", "active-btn"];
+	return (
+		<div className="edit-sidebar edit-section">
+			<CustomButton
+				btnText="Content"
+				onClick={() => onTabChange("content")}
+				classList={
+					tabType === "content"
+						? btnClassList
+						: btnClassList.slice(0, 2)
+				}
+				iconKeyword="toc"
+			/>
+			<CustomButton
+				btnText="Customize"
+				onClick={() => onTabChange("customize")}
+				classList={
+					tabType === "customize"
+						? btnClassList
+						: btnClassList.slice(0, 2)
+				}
+				iconKeyword="edit"
+			/>
+		</div>
+	);
+}
+EditSideBar.propTypes = {
+	tabType: PropTypes.string,
+	onTabChange: PropTypes.func,
+};
+
+export default EditSideBar;
