+++ src/components/EditPanel/ResumeActions.jsx
@@ -1,0 +1,43 @@
+import PropTypes from "prop-types";
+import CustomButton from "../CustomButton";
+
+// For critical resume interactions that involve the entire resume
+function ResumeActions({ loadExampleResume, downloadResume, clearResume }) {
+	return (
+		<div className="resume-actions edit-section">
+			<header className="edit-section-header">
+				<h2>Resume Actions</h2>
+			</header>
+			<section className="edit-section-body">
+				<div className="action-btns-container">
+					<CustomButton
+						btnText="Load Example Resume"
+						classList={["load-example-btn", "button-shrink"]}
+						onClick={loadExampleResume}
+					/>
+
+					<CustomButton
+						btnText="Download"
+						classList={["download-resume-btn", "button-shrink"]}
+						iconKeyword="download"
+						onClick={downloadResume}
+					/>
+
+					<CustomButton
+						btnText="Clear Resume"
+						classList={["clear-resume-btn", "button-shrink"]}
+						onClick={clearResume}
+						iconKeyword="delete"
+					/>
+				</div>
+			</section>
+		</div>
+	);
+}
+ResumeActions.propTypes = {
+	loadExampleResume: PropTypes.func,
+	downloadResume: PropTypes.func,
+	clearResume: PropTypes.func,
+};
+
+export default ResumeActions;
