+++ src/styles/AddItem.css
@@ -1,0 +1,16 @@
+/* 
++ Credits for css-grid transition trick: https://www.youtube.com/watch?v=B_n4YONte5A 
+*/
+.item-section-body {
+	display: grid;
+	grid-template-rows: 0fr;
+	transition: grid-template-rows var(--speed);
+}
+
+.flow-div {
+	overflow: hidden;
+}
+
+.item-section-body.open {
+	grid-template-rows: 1fr;
+}
