+++ src/styles/FormButtons.css
@@ -1,31 +1,3 @@
-.input-group {
-	display: flex;
-	flex-direction: column;
-	gap: 6px;
-}
-.label-text {
-	font-size: 18px;
-	font-weight: 500;
-}
-
-.form-fields {
-	display: flex;
-	flex-direction: column;
-	row-gap: 12px;
-}
-
-.input-group input {
-	background-color: var(--main-bg-clr);
-	border-radius: 12px;
-	padding: 10px;
-	border: 2px solid transparent;
-	outline: none;
-	transition: var(--transition);
-}
-.input-group input:focus {
-	border-color: var(--focus-input-clr);
-}
-
 .form-btn-container {
 	margin-top: 1em;
 	display: flex;
